fix: timer destroying issue

+++ packages/framework/src/plugins/TimerPlugin.ts
@@ -298,6 +298,7 @@ export class Timer {
    * A destroyed timer cannot be restarted.
    */
   public destroy(): void {
+    if (this.isDestroyed) return;
     this._destroy();
     this.plugin.destroyTimer(this);
   }
